move configuration into module
