+++ src/pysonar_scanner/configuration/__init__.py
@@ -1,0 +1,35 @@
+#
+# Sonar Scanner Python
+# Copyright (C) 2011-2024 SonarSource SA.
+# mailto:info AT sonarsource DOT com
+#
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU Lesser General Public
+# License as published by the Free Software Foundation; either
+# version 3 of the License, or (at your option) any later version.
+# This program is distributed in the hope that it will be useful,
+#
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+# Lesser General Public License for more details.
+#
+# You should have received a copy of the GNU Lesser General Public License
+# along with this program; if not, write to the Free Software Foundation,
+# Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+#
+
+from pysonar_scanner.configuration import properties
+from pysonar_scanner.configuration.cli import CliConfigurationLoader
+from pysonar_scanner.utils import filter_none_values
+
+
+def get_static_default_properties() -> dict[str, any]:
+    return {prop.name: prop.default_value for prop in properties.PROPERTIES if prop.default_value is not None}
+
+
+class ConfigurationLoader:
+    def load(self) -> dict[str, any]:
+        return {
+            **get_static_default_properties(),
+            **filter_none_values(CliConfigurationLoader.load()),
+        }
